requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
asyncpg==0.29.0
httpx==0.27.0
tortoise-orm==0.21.3
aerich==0.7.2
celery==5.4.0
redis==5.0.5
eventlet==0.36.1
asgi-lifespan==2.1.0

# 用户单侧
# testcontainers[postgresql]==4.7.1
pytest-asyncio==0.23.7
pytest==8.2.2
pytest-cov==5.0.0
pytest-xdist==3.6.1